requirements.txt
click==8.1.3
parse==1.19.0
python-dotenv==0.21.0
pytimers==3.1

# test dependencies
black
flake8
flake8-import-order
flake8-print
flake8-simplify
mypy
pytest
pytest-cov
